Add support for predefined adressbooks
A new configuration directive `carddav_default_servers` can be used to specify
a list of predefined addressbooks. It supports some replacement strings in
order to work with CardDAV servers that share authentication credentials with
the IMAP server.

In order to spot new (or deleted) predefined addressbooks, we add a new
`default_server` flag to the database. This way we can ensure that we will
never automatically remove user defined calendars.

We also hide and prevent modifications to predefined addressbooks from the
users.

Closes: https://github.com/graviox/Roundcube-CardDAV/issues/4